Add basic benchmarking for simple loads with fwd_pointers
This adds a simple benchmark with pytest-benchmark with writing
a largeish (400 * 400) dataframe of random floats for multiple
symbols.
Then it appends 1 row for each symbol for 10 days followed by
reading all. The idea is to replicate a flow that's closer to
use cases eg. here it's once a day data.

Will add a similar test for one minute data like load as well.

The code has some duplications, but I want to push the benchmarks
else it just keeps growing locally.